move step test client
