+++ tests/test_api.py
@@ -1,14 +1,11 @@
 import json
 import logging
 import uuid
-from datetime import datetime, timedelta, timezone
 from urllib.parse import urljoin
 
-from cryptography import x509
-from cryptography.hazmat.primitives import hashes, serialization
+from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import ec
 from cryptography.hazmat.primitives.serialization import load_pem_public_key
-from cryptography.x509.oid import NameOID
 from fastapi.testclient import TestClient
 from jwcrypto.jwk import JWK
 from jwcrypto.jws import JWS
@@ -16,57 +13,12 @@ from pydantic_settings import SettingsConfigDict
 
 from nodeman.server import NodemanServer
 from nodeman.settings import Settings
-from nodeman.step import StepSignResponse
 from nodeman.utils import generate_x509_csr, jwk_to_alg
+from tests.utils import TestStepClient
 
 ADMIN_TEST_NODE_COUNT = 100
 
 Settings.model_config = SettingsConfigDict(toml_file="tests/test.toml")
-
-
-class TestStepClient:
-    def __init__(self):
-        self.ca_name = "ca.example.com"
-        self.ca_url = "https://ca.example.com"
-
-    def sign_csr(self, csr: x509.CertificateSigningRequest, name: str) -> StepSignResponse:
-        now = datetime.now(tz=timezone.utc)
-        one_day = timedelta(days=1)
-
-        ca_private_key = ec.generate_private_key(ec.SECP256R1())
-        builder = x509.CertificateBuilder()
-        builder = builder.subject_name(x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, self.ca_name)]))
-        builder = builder.issuer_name(x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, self.ca_name)]))
-        builder = builder.not_valid_before(now)
-        builder = builder.not_valid_after(now + one_day)
-        builder = builder.serial_number(x509.random_serial_number())
-        builder = builder.public_key(ca_private_key.public_key())
-        builder = builder.add_extension(
-            x509.BasicConstraints(ca=True, path_length=None),
-            critical=True,
-        )
-        ca_certificate = builder.sign(
-            private_key=ca_private_key,
-            algorithm=hashes.SHA256(),
-        )
-
-        builder = x509.CertificateBuilder()
-        builder = builder.subject_name(x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, name)]))
-        builder = builder.issuer_name(x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, self.ca_name)]))
-        builder = builder.not_valid_before(now)
-        builder = builder.not_valid_after(now + one_day)
-        builder = builder.serial_number(x509.random_serial_number())
-        builder = builder.public_key(csr.public_key())
-        builder = builder.add_extension(x509.SubjectAlternativeName([x509.DNSName(name)]), critical=False)
-        builder = builder.add_extension(
-            x509.BasicConstraints(ca=True, path_length=None),
-            critical=False,
-        )
-        certificate = builder.sign(
-            private_key=ca_private_key,
-            algorithm=hashes.SHA256(),
-        )
-        return StepSignResponse(cert_chain=[certificate], ca_cert=ca_certificate)
 
 
 def get_test_client() -> TestClient:
